libide-tree: import libide-tree from Builder
This is a changed version of libide-tree from Builder so that it can work
outside of the Builder codebase. Mostly around dropping "addins" and
some macros, etc.

This should let us have functionality for building a tree similar to how
the manuals panel is done in Builder which is extremely convenient when
combined with DexFuture.