sdk-dialog: reverse sort by version number
Not exactly ideal way to do this, but works well enough for the SDKs we
care about. GNOME is already pre-sorted higher, and so this will end up
placing the Nightly before 48 and 48 before 47, etc.

When redoing this on libfoundry, we would want to use the version info
from the SDK object for sorting.

Fixes: #13