libide-tree: reset click gesture after claiming
This ensures we keep getting a n_press == 1 for subsequent clicks.

Note that events still come in somewhat strange with this. We get a

 * Press (n_press=1)
 * Release (n_press-1) (this is where we activate the action because there
   is a very long tradition of doing things on release, not press).
 * Reset controller
 * Press (n_press=1)
 * No matching release

The practical result of this is that things work pretty similar for both
single and fast double clicks (being just to expand) not show/hide.

There may be value in adding gtk_gesture_click_reset_counter() to clear
the counter value back to 0 without breaking the sequence state, but I'm
not sure that is worth the effort here.

Fixes: #10