foundry: port Manuals to libfoundry-1
Now that libfoundry has all of the abstractions that originated for Manuals
(and Builder) we can rewrite manuals to be much smaller and defer that
work to libfoundry.

libfoundry-1 should have a 1.0 ready for GNOME 49 to target with this
release. Technically, we could avoid such a large dependency in libfoundry
but that is a massive amount of duplicated code when we'd likely be able
to have libfoundry available anyway.

It also ensures that Builder and libfoundry can share access to a
database of documentation without depending on each other (or having yet
a third copy of the code like there is today).

Libfoundry can handle re-indexing documentation when there are changes to
the available documentation roots, so this solves that issue too.

Lastly, new iconography was used in the tree browsing to help convey what
could be expanded (and when it is expanded).

Resolves: #1
Resolves: #11
Resolves: #14